requirements.txt
flask==3.1.2
flask-cors==6.0.1
httpx==0.28.1
pillow==11.3.0
requests==2.32.5
beautifulsoup4==4.13.5
colorama==0.4.6
anthropic==0.66.0
playwright==1.55.0
python-dotenv==1.1.1
asyncio==4.0.0